refactor: created cogshop config package
