+++ cognite/powerops/resync/config_classes/cogshop/shop_file_config.py
