+++ cognite/powerops/resync/config_classes/cogshop/shop_output_definition.py
