+++ cognite/powerops/resync/config_classes/core/connections.py
